IndexedDB/idbcursor_advance_index2.htm

<!DOCTYPE html>
<meta charset="utf-8">
<title>IDBCursor.advance() - index - errors in count argument</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://dvcs.w3.org/hg/IndexedDB/raw-file/tip/Overview.html#widl-IDBCursor-advance-void-unsigned-long-count">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support.js"></script>
<div id="log"></div>
<script>
    var db,
        t = async_test(),
        records = [{ pKey: "primaryKey_0", iKey: "indexKey_0" },
                   { pKey: "primaryKey_1", iKey: "indexKey_1" }];

    var open_rq = createdb(t);
    open_rq.onupgradeneeded = function (event) {
        db = event.target.result;
        var objStore = db.createObjectStore("store", {keyPath : "pKey"});
        objStore.createIndex("index", "iKey");
        for (var i = 0; i < records.length; i++) {
            objStore.add(records[i]);
        }
        var index = objStore.index("index");
        var rqCursor = index.openCursor();
        rqCursor.onsuccess = t.step_func(function(event) {
            var cursor = event.target.result;
            assert_cursor_exists(cursor);

            assert_throws(new TypeError(), function() {
                cursor.advance(0);
            }, "Calling advance() with count argument 0 should throw TypeError.");

            assert_throws("InvalidStateError", function() {
                cursor.advance(1);
                cursor.advance(1);
            }, "Calling advance() twice should throw DOMException when the cursor is currently being iterated.");

            t.done();
        });
    }
</script>
